fix: audit third pass — controller split, React.memo, loading skeletons, WS status dot
#43 Controller split:
- controller.ts (1,056 lines) → 36-line barrel re-export
- controller-sync.ts (321 lines): mobileSyncOnce, refreshInbox, loadHistory, loadReviewPacket, loadReviewFile
- controller-compose.ts (385 lines): attachments, enhance, steer, resume, correction draft
- controller-actions.ts (339 lines): runAction, disposition, copy, refresh, focus, stop, diff
- Zero import changes needed — barrel re-exports preserve all consumers

#43 React.memo:
- SquadRail wrapped in memo() — skips re-render when session list unchanged
- ComposeBar wrapped in memo() — skips re-render when draft/state unchanged
- These are the two highest-frequency re-render surfaces

#45 Loading skeletons:
- ShimmerCard component: pulse animation, Apple-style rounded card
- All 5 lazy-loaded panels get loading fallback (no more pop-in)

#48 WS connection status:
- Green/amber/red dot in TopBar header label
- Green = connected, amber = connecting/reconnecting, red = disconnected
- Inline styles, 6px dot, title tooltip with state name

Apple PWA meta tags:
- Already present via Next.js metadata API (appleWebApp.capable + statusBarStyle)
- No changes needed
